Implement the Win64 EH prolog instruction methods on the base MCStreamer.
I had to change the API slightly to avoid overloading issues.

llvm-svn: 131666

+++ llvm/include/llvm/MC/MCStreamer.h
@@ -470,11 +470,11 @@ namespace llvm {
     virtual void EmitWin64EHHandler(const MCSymbol *Sym, bool Unwind,
                                     bool Except);
     virtual void EmitWin64EHHandlerData();
-    virtual void EmitWin64EHPushReg(int64_t Register);
-    virtual void EmitWin64EHSetFrame(int64_t Register, int64_t Offset);
-    virtual void EmitWin64EHAllocStack(int64_t Size);
-    virtual void EmitWin64EHSaveReg(int64_t Register, int64_t Offset);
-    virtual void EmitWin64EHSaveXMM(int64_t Register, int64_t Offset);
+    virtual void EmitWin64EHPushReg(unsigned Register);
+    virtual void EmitWin64EHSetFrame(unsigned Register, unsigned Offset);
+    virtual void EmitWin64EHAllocStack(unsigned Size);
+    virtual void EmitWin64EHSaveReg(unsigned Register, unsigned Offset);
+    virtual void EmitWin64EHSaveXMM(unsigned Register, unsigned Offset);
     virtual void EmitWin64EHPushFrame(bool Code);
     virtual void EmitWin64EHEndProlog();
